start work towards enabling launching fuzzing within docker

+++ src/ci/azcopy.sh
@@ -5,12 +5,15 @@
 
 set -ex
 
-mkdir -p artifacts/azcopy
-
-wget -O azcopy.zip https://aka.ms/downloadazcopy-v10-windows
-unzip azcopy.zip
-mv azcopy_windows*/* artifacts/azcopy/
-
-wget -O azcopy.tgz https://aka.ms/downloadazcopy-v10-linux
-tar zxvf azcopy.tgz
-mv azcopy_linux_amd64*/* artifacts/azcopy/
+if [ ! -f artifacts/azcopy/azcopy ]; then
+    mkdir -p artifacts/azcopy
+    
+    wget -O azcopy.zip https://aka.ms/downloadazcopy-v10-windows
+    unzip azcopy.zip
+    mv azcopy_windows*/* artifacts/azcopy/
+    
+    wget -O azcopy.tgz https://aka.ms/downloadazcopy-v10-linux
+    tar zxvf azcopy.tgz
+    mv azcopy_linux_amd64*/* artifacts/azcopy/
+    rm -r azcopy*
+fi
